Add muzak-rebase-index.
muzak-rebase-index can be used to "rebase" your music index to a new
music tree, as often happens when moving a music library between
systems or drives. It's useful for avoiding an expensive reindexing operation
if the only thing that's changed is the path of the music tree.

Example: `muzak-rebase-index /new/root --force` would replace all
occurrences of Muzak::Config.music with "/new/root" in the index,
and update Config.music itself to reflect the new tree.